+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.3.26] - 2023-05-24
+### Added
+- Hue and Cry! button added to combat tracker.
+- Foundry v11 api support.
+- Combatants who are ready to act have their inits highlighted.
+### Changed
+- Combatants now sorted by NPC status and name, rather than initiative.
+- Removed extraneous details from combat tracker config.
+- Reach animation tweaks.
+
 ## [0.3.25] - 2023-05-21
 ### Added
 - Chat card visual improvements.